New EndPoint "coudProviderInfo"
