+++ go.mod
@@ -19,6 +19,7 @@ require (
 )
 
 require (
+	github.com/armosec/utils-go v0.0.12
 	github.com/kr/text v0.2.0 // indirect
 	github.com/niemeyer/pretty v0.0.0-20200227124842-a10e7caefd8e // indirect
 	github.com/pkg/errors v0.9.1 // indirect
